Delete Modal, rename Sidebar to Modal
