+++ src/Modal/Modal.stories.js
@@ -1,28 +1,77 @@
 import Modal from "./Modal.vue";
+import ShowcaseCenter from "./ShowcaseCenter.vue"
+import { withKnobs, text, select, boolean } from '@storybook/addon-knobs';
+import data from "./data"
 
 export default {
   title: "Modal",
-  component: Modal
+  component: Modal,
+  decorators: [withKnobs]
 };
 
 export const normal = () => ({
   components: { Modal },
   data: function () {
     return {
-      modal: null,
-      lorem:
-        "It is a long established fact that a reader will be distracted by the readable content of a page when looking at its layout. The point of using Lorem Ipsum is that it has a more-or-less normal distribution of letters, as opposed to using 'Content here, content here', making it look like readable English. Many desktop publishing packages and web page editors now use Lorem Ipsum as their default model text, and a search for 'lorem ipsum' will uncover many web sites still in their infancy. Various versions have evolved over the years, sometimes by accident, sometimes on purpose (injected humour and the like). Where does it come from? Contrary to popular belief, Lorem Ipsum is not simply random text. It has roots in a piece of classical Latin literature from 45 BC, making it over 2000 years old. Richard McClintock, a Latin professor at Hampden-Sydney College in Virginia, looked up one of the more obscure Latin words, consectetur, from a Lorem Ipsum passage, and going through the cites of the word in classical literature, discovered the undoubtable source. Lorem Ipsum comes from sections 1.10.32 and 1.10.33 of de Finibus Bonorum et Malorum (The Extremes of Good and Evil) by Cicero, written in 45 BC. This book is a treatise on the theory of ethics, very popular during the Renaissance. The first line of Lorem Ipsum, Lorem ipsum dolor sit amet.., comes from a line in section 1.10.32. The standard chunk of Lorem Ipsum used since the 1500s is reproduced below for those interested. Sections 1.10.32 and 1.10.33 from de Finibus Bonorum et Malorum by Cicero are also reproduced in their exact original form, accompanied by English versions from the 1914 translation by H. Rackham. Where can I get some? There are many variations of passages of Lorem Ipsum available, but the majority have suffered alteration in some form, by injected humour, or randomised words which don't look even slightly believable. If you are going to use a passage of Lorem Ipsum, you need to be sure there isn't anything embarrassing hidden in the middle of text. All the Lorem Ipsum generators on the Internet tend to repeat predefined chunks as necessary, making this the first true generator on the Internet. It uses a dictionary of over 200 Latin words, combined with a handful of model sentence structures, to generate Lorem Ipsum which looks reasonable. The generated Lorem Ipsum is therefore always free from repetition, injected humour, or non-characteristic words etc. "
-    };
+      visible: null,
+      data: data.sidebar,
+    }
+  },
+  props: {
+    side: {
+      default: select(
+        "Side",
+        ["left", "right", "bottom", "center"],
+        "left"
+      )
+    },
+    width: {
+      default: text("Width", "")
+    },
+    maxWidth: {
+      default: text("Max width", "")
+    },
+    height: {
+      default: text("Height", "")
+    },
+    maxHeight: {
+      default: text("Max height", "")
+    },
+    marginTop: {
+      default: text("Margin top", "")
+    },
+    backgroundColor: {
+      default: text("Background color", "rgba(0, 0, 0, 0.35)")
+    },
+    boxShadow: {
+      default: text("Box shadow", "none")
+    },
+    fullscreen: {
+      default: boolean("Fullscreen", false)
+    },
+    sidebarContent: {
+      default: text("Sheet content", data.sidebar.lorem.join(""))
+    },
   },
   template: `
     <div>
-      <button @click="modal = !modal">Open modal</button>
-      <modal :visible="modal" @input="modal = false">
-        <div class="content" style="padding: 2rem">
-          <div>{{lorem}}</div>
-          <div><input type="text"></div>
+      <div>
+        <Modal v-bind="{side, width, maxWidth, fullscreen, height, maxHeight, marginTop, backgroundColor, boxShadow}" :visible="visible" v-if="visible" @visible="visible = $event">
+          <div>{{sidebarContent}}</div>
+        </Modal>
+        <div>
+          <p v-for="text in data.lorem.slice(1,4)">{{text}}</p>
+          <button @click="visible = !visible">Open sidebar</button>
+          <p v-for="text in data.lorem">{{text}}</p>
         </div>
-      </modal>
+      </div>
     </div>
   `
 });
+
+export const center = () => ({
+  components: { ShowcaseCenter },
+  template: `
+    <showcase-center/>
+  `
+})
