+++ src/Modal/Modal.vue
@@ -1,280 +1,436 @@
 <template>
   <div>
-    <div class="wrapper" v-if="visible">
-      <transition name="overlay" @after-leave="close" appear>
-        <div class="overlay" v-if="visible && show" :style="{'--overlay-background': overlay}"></div>
-      </transition>
-      <div class="modal-wrapper" @click.self="show = false">
-        <transition name="modal" appear>
-          <div class="modal" v-if="visible && show" tabindex="0" @keydown.esc="show = false" ref="modal">
-            <button
-              :class="['icon-cross', {'icon-cross__dark': button == 'dark'}]"
-              @click="show = false"
-            >
-              <svg width="24" height="24" fill="none" xmlns="http://www.w3.org/2000/svg">
-                <path d="M4 4l16 16m0-16L4 20" stroke-width="1.5" stroke-linecap="round" />
-              </svg>
-            </button>
-            <!-- @slot Contents of the modal. -->
-            <slot></slot>
-            <button
-              :class="['icon-cross', {'icon-cross__dark': button == 'dark'}]"
-              @click="show = false"
-            >
-              <svg width="24" height="24" fill="none" xmlns="http://www.w3.org/2000/svg">
-                <path d="M4 4l16 16m0-16L4 20" stroke-width="1.5" stroke-linecap="round" />
-              </svg>
-            </button>
-          </div>
-        </transition>
+    <transition name="overlay" appear>
+      <div class="overlay"
+           ref="overlay"
+           :style="{'background-color': backgroundColor || 'rgba(0, 0, 0, 0.35)'}"
+           v-if="visible && visibleLocal"
+           @click="close"
+           @touchstart="touchstart"
+           @touchmove="touchmove"
+           @touchend="touchend">
       </div>
-    </div>
+    </transition>
+    <transition :name="`sidebar__${side}`" @after-leave="emitVisible()" appear>
+      <div :class="['sidebar', `sidebar__side__${side}`]"
+           ref="sidebar"
+           v-if="visible && visibleLocal"
+           :style="style"
+           @click.self="sidebarClick"
+           @touchstart="touchstart"
+           @touchmove="touchmove"
+           @touchend="touchend">
+        <!-- @slot Contents of the sidebar. -->
+        <div @scroll="detectScrolling" ref="content" :class="[`sidebar__content`, `sidebar__content__side__${side}`, `sidebar__fullscreen__${!!(fullscreenComputed)}`]">
+          <slot/>
+        </div>
+        <div class="close" v-if="side === 'center' && buttonClose" @click="close">
+          <svg class="close__icon" width="100%" height="100%" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+            <path d="M4 4l16 16m0-16L4 20" stroke-width="1.5" stroke-linecap="round" />
+          </svg>
+        </div>
+      </div>
+    </transition>
   </div>
 </template>
 
+<style scoped>
+.overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  z-index: 100000;
+  height: 100vh;
+  width: 100vw;
+}
+.sidebar {
+  position: fixed;
+  overflow-y: hidden;
+  z-index: 100000;
+  transform: translateX(var(--sidebar-translate-x)) translateY(var(--sidebar-translate-y));
+  -webkit-overflow-scrolling: touch;
+}
+.close {
+  border-radius: 50%;
+  stroke: rgba(255,255,255,.75);
+  box-sizing: border-box;
+  padding: 8px;
+  width: 48px;
+  height: 48px;
+  position: absolute;
+  top: 0;
+  right: 0;
+  margin: 1rem;
+}
+.close__icon {
+  display: block;
+}
+.sidebar.sidebar__side__left {
+  top: 0;
+  left: 0;
+  right: initial;
+  width: var(--sidebar-width, 300px);
+  max-width: var(--sidebar-width, 75%);
+  height: var(--sidebar-height, 100%);
+  max-height: var(--sidebar-max-height, 100%);
+  box-shadow: var(--sidebar-box-shadow);
+}
+.sidebar.sidebar__side__right {
+  top: 0;
+  left: initial;
+  right: 0;
+  width: var(--sidebar-width, 300px);
+  max-width: var(--sidebar-width, 75%);
+  height: var(--sidebar-height, 100%);
+  max-height: var(--sidebar-max-height, 100%);
+  box-shadow: var(--sidebar-box-shadow);
+}
+.sidebar.sidebar__side__bottom {
+  top: 0;
+  left: 0;
+  right: 0;
+  max-width: initial;
+  overflow-y: scroll;
+  -webkit-overflow-scrolling: touch;
+  width: var(--sidebar-width, 100%);
+  max-width: var(--sidebar-width, 100%);
+  height: 100%;
+  margin-left: auto;
+  margin-right: auto;
+}
+.sidebar.sidebar__side__center {
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  width: 100%;
+  max-width: 100%;
+  overflow-y: scroll;
+  -webkit-overflow-scrolling: touch;
+  pointer-events: all;
+  height: 100vh;
+}
+.sidebar__content {
+  background: white;
+  position: absolute;
+  pointer-events: all;
+  overflow-y: scroll;
+  -webkit-overflow-scrolling: touch;
+  width: 100%;
+  height: 100%;
+}
+.sidebar__content.sidebar__content__side__bottom {
+  margin-top: var(--sidebar-top);
+  overflow-y: hidden;
+  height: auto;
+  box-shadow: var(--sidebar-box-shadow);
+}
+
+.sidebar__content.sidebar__content__side__center {
+  pointer-events: all;
+  position: absolute;
+  width: var(--sidebar-width, 600px);
+  max-width: var(--sidebar-max-width, 90%);
+  height: var(--sidebar-height, auto);
+  max-height: var(--sidebar-max-height, none);
+  top: var(--sidebar-top);
+  transform: translateX(-50%);
+  left: 50%;
+  border-radius: var(--sidebar-border-radius);
+  box-shadow: var(--sidebar-box-shadow);
+  margin-bottom: 20px;
+  border-radius: .5rem;
+}
+
+.sidebar__content.sidebar__content__side__center.sidebar__fullscreen__true {
+  top: 0;
+  width: 100%;
+  height: 100%;
+  max-width: 100%;
+  max-height: 100%;
+  margin-bottom: initial;
+  border-radius: 0;
+}
+
+.overlay-enter-active {
+  transition: all .25s ease-out;
+}
+.overlay-enter {
+  opacity: 0;
+}
+.overlay-enter-to {
+  opacity: 1;
+}
+.overlay-leave-active {
+  transition: all .25s;
+}
+.overlay-leave {
+  opacity: 1;
+}
+.overlay-leave-to {
+  opacity: 0;
+}
+.sidebar__left-enter-active,
+.sidebar__right-enter-active,
+.sidebar__bottom-enter-active,
+.sidebar__center-enter-active,
+.sidebar__left-leave-active,
+.sidebar__right-leave-active,
+.sidebar__bottom-leave-active,
+.sidebar__center-leave-active {
+  transition: all .5s;
+}
+.sidebar__left-enter,
+.sidebar__left-leave-to {
+  transform: translateX(-100%);
+}
+.sidebar__right-enter,
+.sidebar__right-leave-to {
+  transform: translateX(100%)
+}
+.sidebar__left-enter-to,
+.sidebar__left-leave
+.sidebar__right-enter-to,
+.sidebar__right-leave {
+  transform: translateX(0)
+}
+.sidebar__bottom-enter,
+.sidebar__bottom-leave-to {
+  transform: translateY(100%)
+}
+.sidebar__bottom-enter-to,
+.sidebar__bottom-leave {
+  transform: translateY(0)
+}
+.sidebar__center-enter, .sidebar__center-leave-to {
+  opacity: 0;
+  transform: scale(.95)
+}
+.sidebar__center-enter-to, .sidebar__center-leave {
+  opacity: 1;
+  transform: scale(1);
+}
+</style>
+
 <script>
 /**
- * The `Modal` component opens a modal window that displays `default` slot content.
- * Clicking on the close button or the overlay sends an `input` event with the value
- * `true`.
- * 
- * `<modal visible="!!(isVisible)" @input="isVisible = $event">`
+ * `Sidebar` is a sheet that transitions on top of the main content and allows
+ * displaying auxiliary information, such as table of contents or global navigation.
  */
 export default {
   props: {
     /**
-     * Toggles visibility of the modal.
+     * Toggles visibility of the component.
      */
     visible: {
       type: Boolean,
-      default: true
+      default: false
     },
     /**
-     * Color of the overlay.
+     * Width of the sidebar.
      */
-    overlay: {
+    width: {
       type: String,
-      default: "rgba(0, 0, 0, 0.25)"
     },
     /**
-     * Color of the close button: `light` | `dark`
+     * Maximum width of the sidebar.
      */
-    button: {
+    maxWidth: {
       type: String,
-      default: "light"
+    },
+    /**
+     * Height of the sidebar.
+     */
+    height: {
+      type: String,
+    },
+    /**
+     * Maximum height of the sidebar.
+     */
+    maxHeight: {
+      type: String,
+    },
+    /**
+     * `left` | `right` | `bottom`
+     */
+    side: {
+      type: String,
+      default: "left"
+    },
+    /**
+     * CSS `background-color` of the overlay.
+     */
+    backgroundColor: {
+      type: String,
+      default: 'rgba(0, 0, 0, 0.35)'
+    },
+    /**
+     * CSS `box-shadow` of the sidebar sheet.
+     */
+    boxShadow: {
+      type: String,
+      default: "none"
+    },
+    /**
+     * Vertical height of overlay
+     */
+    marginTop: {
+      type: Number
+    },
+    /**
+     * Go fullscreen when viewport is narrower than width
+     */
+    fullscreen: {
+      type: Boolean,
+      default: false
+    },
+    /**
+     * Add default close button for centered modal
+     */
+    buttonClose: {
+      type: Boolean,
+      default: false
     }
   },
   data: function() {
     return {
-      show: true
+      visibleLocal: true,
+      startX: null,
+      startY: null,
+      currentX: null,
+      currentY: null,
+      translateX: null,
+      translateY: null,
+      isScrolling: null,
+      sheetTop: null,
+      fullscreenComputed: null,
     };
   },
   watch: {
-    visible(becomesVisible) {
-      if (becomesVisible) {
-        this.$nextTick(() => {
-          if (this.$refs.modal) this.$refs.modal.focus()
-        })
+    visible(newValue, oldValue) {
+      if (newValue) {
+        const sidebar = this.$refs.sidebar;
+        if (sidebar) {
+          sidebar.addEventListener("transitionend", () => {
+            sidebar.style.transition = "";
+          });
+        }
+        this.visibleLocal = true;
       }
     }
   },
+  computed: {
+    deltaX() {
+      return this.currentX - this.startX
+    },
+    deltaY() {
+      return this.currentY - this.startY
+    },
+    style() {
+      return {
+        "--sidebar-max-width": this.maxWidth,
+        "--sidebar-width": this.width,
+        "--sidebar-max-height": this.maxHeight,
+        "--sidebar-height": this.height,
+        "--sidebar-box-shadow": this.boxShadow,
+        "--sidebar-top": this.sheetTop + 'px',
+        "--sidebar-translate-x": `${this.translateX || 0}px`,
+        "--sidebar-translate-y": `${this.translateY || 0}px`
+      };
+    }
+  },
+  mounted() {
+    this.onResize()
+    document.querySelector("body").style.overflow = "hidden"
+    window.addEventListener("resize", this.onResize)
+  },
   methods: {
-    close() {
-      this.show = true;
+    onResize() {
+      this.sheetCenterPosition()
+      if (this.side === "center" && this.$refs.content) {
+        this.fullscreenComputed = (window.innerWidth <= parseInt(this.width) && this.fullscreen)
+      }
+    },
+    sheetCenterPosition() {
+      this.sheetCenterPosition
+      if (this.side === "center" && this.$refs.content) {
+        const
+          content = this.$refs.content.offsetHeight,
+          height = window.innerHeight
+        this.sheetTop = content > height - 40 ? 20 : (height - content) / 2
+      }
+      if (this.side === "bottom" && this.$refs.content) {
+        const
+          content = this.$refs.content.offsetHeight,
+          height = window.innerHeight
+        this.sheetTop = content > height - (this.marginTop || 100)
+          ? (this.marginTop || 100)
+          : height - content
+      }
+    },
+    sidebarClick(e) {
+      if (this.side === "center") this.visibleLocal = null;
+      if (this.side === 'bottom') this.visibleLocal = null
+    },
+    detectScrolling(e) {
+      this.isScrolling = true
+    },
+    emitVisible() {
+      document.querySelector("body").style.overflow = ""
       /**
-       * Sends `true` when closing the modal.
+       * Sends `false` when closing the sidebar.
        * @type {Event}
        */
-      this.$emit("input", true);
-    }
+      this.$emit('visible', false)
+    },
+    close(e) {
+      this.visibleLocal = null;
+      if (this.$refs["overlay"]) {
+        this.$refs["overlay"].style["pointer-events"] = "none";
+        if (e.clientX && e.clientY) {
+          const doc = document.elementFromPoint(e.clientX, e.clientY);
+          if (doc && doc.click) doc.click();
+        }
+      }
+    },
+    touchstart(e) {
+      const sidebar = this.$refs.sidebar
+      const content = this.$refs.content
+      if (sidebar) sidebar.style.transition = ""
+      this.currentX = this.startX = e.changedTouches[0].clientX;
+      this.currentY = this.startY = e.changedTouches[0].clientY;
+    },
+    touchmove(e) {
+      this.currentX = e.changedTouches[0].clientX;
+      this.currentY = e.changedTouches[0].clientY;
+      if (this.side === "left" && !this.isScrolling) {
+        this.translateX = this.deltaX > 0 ? 0 : this.deltaX
+      }
+      if (this.side === "right" && !this.isScrolling) {
+        this.translateX = this.deltaX < 0 ? 0 : this.deltaX
+      }
+    },
+    touchend(e) {
+      const
+        overThresholdX = Math.abs(this.deltaX * 100 / window.screen.width) > 25
+        sidebar = this.$refs.sidebar
+      if (this.side === "left") {
+        this.translateX = this.deltaX > 0 ? 0 : this.deltaX
+      } else if (this.side === "right") {
+        this.translateX = this.deltaX < 0 ? 0 : this.deltaX
+      }
+      if (overThresholdX && !this.isScrolling && (this.side === "left" || this.side === "right")) {
+        this.close(e)
+      } else {
+        if (sidebar) sidebar.style.transition = "all .5s";
+      }
+      this.startX = null;
+      this.startY = null;
+      this.currentX = null;
+      this.currentY = null;
+      this.translateX = null;
+      this.isScrolling = null;
+    },
   }
 };
-</script>
-
-<style scoped>
-.wrapper {
-  position: fixed;
-  z-index: 5000;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  overflow: scroll;
-  -webkit-overflow-scrolling: touch;
-}
-
-.overlay {
-  background: var(--overlay-background);
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  pointer-events: all;
-}
-
-.modal-wrapper {
-  display: grid;
-  justify-content: center;
-  align-items: center;
-  grid-template-columns: 1fr;
-  height: 100%;
-  position: relative;
-  overflow-x: hidden;
-  padding-right: 4rem;
-  padding-left: 4rem;
-}
-
-.modal {
-  justify-self: center;
-  background: #fff;
-  border-radius: 8px;
-  position: relative;
-  pointer-events: all;
-  margin-top: 4rem;
-  margin-bottom: 4rem;
-  box-shadow: 0 5px 30px 0 rgba(0, 0, 0, 0.2);
-  outline: none;
-}
-
-.icon-cross {
-  position: absolute;
-  right: -1.5rem;
-  top: -1.5rem;
-  padding: 0.5rem;
-  background: #eff0f5;
-  background-clip: padding-box;
-  border-radius: 1000px;
-  box-shadow: 0px 12px 48px rgba(22, 25, 49, 0.2);
-  border: 1px solid rgba(176, 180, 207, 0.2);
-  outline: none;
-  width: 3rem;
-  height: 3rem;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  z-index: 1000;
-  transition: box-shadow 0.3s, background 0.3s;
-  stroke: #a2a3ad;
-}
-
-.icon-cross.icon-cross__dark {
-  background: #222;
-  box-shadow: 0px 12px 48px rgba(0, 0, 0, 0.75);
-  border: none;
-  stroke: #555;
-}
-
-.icon-cross.icon-cross__dark:hover,
-.icon-cross.icon-cross__dark:focus {
-  background: #333;
-  box-shadow: none;
-  box-shadow: 0px 12px 48px rgba(0, 0, 0, 0.75);
-  border: none;
-  stroke: #777;
-}
-
-.icon-cross:hover,
-.icon-cross:focus {
-  background-color: #fff;
-  box-shadow: 0px 20px 48px rgba(22, 25, 49, 0.3);
-  transition: box-shadow 0.1s, background 0.1s;
-}
-.icon-cross:active {
-  background-color: #eff0f5;
-  transition: none;
-}
-
-.overlay-enter-active {
-  transition: opacity 2s;
-}
-
-.overlay-leave-active {
-  transition: opacity 0.75s;
-}
-
-.overlay-enter {
-  opacity: 0;
-}
-
-.overlay-enter-to {
-  opacity: 1;
-}
-
-.overlay-leave {
-  opacity: 1;
-}
-
-.overlay-leave-to {
-  opacity: 0;
-}
-
-.modal-enter-active {
-  transition: opacity 0.25s, transform 0.75s;
-}
-
-.modal-leave-active {
-  transition: opacity 0.75s, transform 1s;
-}
-
-.modal-enter {
-  opacity: 0;
-  transform: scale(0.95);
-}
-
-.modal-enter-to {
-  opacity: 1;
-  transform: scale(1);
-}
-
-.modal-leave {
-  opacity: 1;
-  transform: scale(1);
-}
-
-.modal-leave-to {
-  opacity: 0;
-  transform: scale(0.95);
-}
-
-@media screen and (max-width: 900px) {
-  .wrapper {
-    padding: 0;
-  }
-  .icon-cross {
-    position: fixed;
-    box-shadow: 0px 12px 48px rgba(22, 25, 49, 0.2);
-    border-radius: 1000px;
-    border: 1px solid #fff;
-    top: 1rem;
-    right: 1rem;
-    background-color: #fff;
-  }
-  .aside {
-    margin-top: 3rem;
-  }
-  .modal {
-    width: 100%;
-    margin: 0;
-    border-radius: 0;
-    min-height: 100vh;
-  }
-
-  .modal-wrapper {
-    display: block;
-    padding-left: 0;
-    padding-right: 0;
-  }
-  .container {
-    display: block;
-    overflow: hidden;
-  }
-  .content {
-    margin-top: 3rem;
-  }
-  .content__paragraph {
-    max-width: initial;
-  }
-}
-</style>
+</script>
