+++ src/index.js
@@ -2,6 +2,5 @@ export { default as Modal } from "./Modal/Modal";
 export { default as Tooltip } from "./Tooltip/Tooltip";
 export { default as PageHeader } from "./PageHeader/PageHeader";
 export { default as CookieBanner } from "./CookieBanner/CookieBanner";
-export { default as Sidebar } from "./Sidebar/Sidebar";
 export { default as Footer } from "./Footer/Footer";
 export { default as CodeBlock } from "./CodeBlock/CodeBlock";
